# gateway-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Book Shop",
        "title": "Car moyka",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/translators": {
            "get": {
                "description": "Get all translators with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translators"
                ],
                "summary": "Get all translators",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Translator Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Translator Surname",
                        "name": "surname",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsGetAllRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a translator's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translators"
                ],
                "summary": "Update a translator",
                "parameters": [
                    {
                        "description": "Translator Update Details",
                        "name": "translator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new translator with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translators"
                ],
                "summary": "Create a new translator",
                "parameters": [
                    {
                        "description": "Translator Details",
                        "name": "translator",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsRes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/translators/{id}": {
            "get": {
                "description": "Get a translator by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translators"
                ],
                "summary": "Get translator by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Translator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.TranslatorsRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a translator by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translators"
                ],
                "summary": "Delete a translator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Translator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/books.Void"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "books.TranslatorsGetAllRes": {
            "type": "object",
            "properties": {
                "translators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/books.TranslatorsRes"
                    }
                }
            }
        },
        "books.TranslatorsRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "books.TranslatorsUpdateReq": {
            "type": "object",
            "properties": {
                "translator": {
                    "$ref": "#/definitions/books.TranslatorsRes"
                }
            }
        },
        "books.Void": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}